+++ lib/src/db.rs
@@ -1,5 +1,6 @@
 //! Persistent, ACID compliant, threadsafe to-disk store.
 //! Powered by Sled - an embedded database.
+//! See [Db]
 
 mod migrations;
 mod prop_val_sub_index;
@@ -16,6 +17,7 @@ use std::{
     vec,
 };
 
+use mail_send::{Connected, Transport};
 use tracing::info;
 use tracing::instrument;
 use trees::{Method, Operation, Transaction, Tree};
@@ -24,11 +26,10 @@ use crate::{
     agents::ForAgent,
     atomic_url::AtomicUrl,
     atoms::IndexAtom,
-    commit::{CommitOpts, CommitResponse},
-    db::{
-        query_index::{requires_query_index, NO_VALUE},
-        val_prop_sub_index::find_in_val_prop_sub_index,
-    },
+    commit::CommitOpts,
+    commit::CommitResponse,
+    db::{query_index::requires_query_index, val_prop_sub_index::find_in_val_prop_sub_index},
+    email::{self, MailMessage},
     endpoints::{default_endpoints, Endpoint, HandleGetContext},
     errors::{AtomicError, AtomicResult},
     resources::PropVals,
@@ -43,7 +44,7 @@ use self::{
     prop_val_sub_index::{add_atom_to_prop_val_sub_index, find_in_prop_val_sub_index},
     query_index::{
         check_if_atom_matches_watched_query_filters, query_sorted_indexed, should_include_resource,
-        update_indexed_member, IndexIterator, QueryFilter,
+        update_indexed_member, IndexIterator, QueryFilter, NO_VALUE,
     },
     val_prop_sub_index::add_atom_to_valpropsub_index,
 };
@@ -55,14 +56,15 @@ type HandleCommit = Box<dyn Fn(&CommitResponse) + Send + Sync>;
 /// The String on the left represents a Property URL, and the second one is the set of subjects.
 pub type PropSubjectMap = HashMap<String, HashSet<String>>;
 
-/// The Db is a persistent on-disk Atomic Data store.
+/// A persistent on-disk Atomic Data store.
 /// It's an implementation of [Storelike].
 /// It uses [sled::Tree]s as Key Value stores.
 /// It stores [Resource]s as [PropVals]s by their subject as key.
 /// It builds a value index for performant [Query]s.
 /// It keeps track of Queries and updates their index when [crate::Commit]s are applied.
-/// You can pass a custom `on_commit` function to run at Commit time.
 /// `Db` should be easily, cheaply clone-able, as users of this library could have one `Db` per connection.
+/// Note that [plugins](crate::plugins) can add their own endpoints to the [Db],
+/// and can use [tokio::spawn] to start concurrent tasks.
 #[derive(Clone)]
 pub struct Db {
     /// The Key-Value store that contains all data.
@@ -88,6 +90,9 @@ pub struct Db {
     on_commit: Option<Arc<HandleCommit>>,
     /// Where the DB is stored on disk.
     path: std::path::PathBuf,
+    handle_commit: Option<Arc<HandleCommit>>,
+    /// Email SMTP client for sending email.
+    smtp_client: Option<Arc<Mutex<Transport<'static, Connected>>>>,
 }
 
 impl Db {
@@ -109,12 +114,14 @@ impl Db {
             default_agent: Arc::new(Mutex::new(None)),
             resources,
             reference_index,
+            on_commit: None,
             query_index,
             prop_val_sub_index,
             server_url: AtomicUrl::try_from(server_url)?,
             watched_queries,
             endpoints: default_endpoints(),
-            on_commit: None,
+            handle_commit: None,
+            smtp_client: None,
         };
         migrate_maybe(&store).map(|e| format!("Error during migration of database: {:?}", e))?;
         crate::populate::populate_base_models(&store)
@@ -218,7 +225,7 @@ impl Db {
     /// Sets a function that is called whenever a [Commit::apply] is called.
     /// This can be used to listen to events.
     pub fn set_handle_commit(&mut self, on_commit: HandleCommit) {
-        self.on_commit = Some(Arc::new(on_commit));
+        self.handle_commit = Some(Arc::new(on_commit));
     }
 
     /// Finds resource by Subject, return PropVals HashMap
@@ -395,14 +402,15 @@ impl Db {
         let atoms = self.get_index_iterator_for_query(q);
 
         for (i, atom_res) in atoms.enumerate() {
-            let atom = atom_res?;
-            if !q.include_external && self.is_external_subject(&atom.subject).unwrap() {
-                continue;
-            }
-
             total_count += 1;
 
             if q.offset > i {
+                continue;
+            }
+
+            let atom = atom_res?;
+
+            if !q.include_external && self.is_external_subject(&atom.subject)? {
                 continue;
             }
 
@@ -475,6 +483,37 @@ impl Db {
             )
             .map_err(|e| format!("Checking atom went wrong: {}", e))?;
         }
+        Ok(())
+    }
+
+    /// Adds an [Endpoint] to the store. This means adding a route with custom behavior.
+    pub fn register_endpoint(&mut self, endpoint: Endpoint) {
+        self.endpoints.push(endpoint);
+    }
+
+    /// Registers an SMTP client to the store, allowing the store to send emails.
+    pub async fn set_smtp_config(
+        &mut self,
+        smtp_config: crate::email::SmtpConfig,
+    ) -> AtomicResult<()> {
+        self.smtp_client = Some(Arc::new(Mutex::new(
+            crate::email::get_smtp_client(smtp_config).await?,
+        )));
+        Ok(())
+    }
+
+    pub async fn send_email(&self, message: MailMessage) -> AtomicResult<()> {
+        let mut client = self
+            .smtp_client
+            .as_ref()
+            .ok_or_else(|| {
+                AtomicError::other_error(
+                    "No SMTP client configured. Please call set_smtp_config first.".into(),
+                )
+            })?
+            .lock()
+            .await;
+        email::send_mail(&mut client, message).await?;
         Ok(())
     }
 }
@@ -803,7 +842,7 @@ impl Storelike for Db {
     }
 
     fn handle_commit(&self, commit_response: &CommitResponse) {
-        if let Some(fun) = &self.on_commit {
+        if let Some(fun) = &self.handle_commit {
             fun(commit_response);
         }
     }